+++ app/boot/themeHandler.ts
