+++ package.json
@@ -25,7 +25,7 @@
   "devDependencies": {
     "@types/chai": "^4.1.6",
     "@types/mocha": "^5.2.7",
-    "@types/node": "^10.14.8",
+    "@types/node": "^10.14.13",
     "babel-eslint": "^8.2.3",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
