+++ test.ts
@@ -1,0 +1,4 @@
+import config from './testconfig';
+
+
+console.log(config('THEME'))
