+++ tsconfig.json
@@ -6,6 +6,9 @@
         "paths": {
             "*": ["./declarations"]
         },
+        "typeRoots": [
+          "./node_modules/@types"
+        ],
         "module": "commonjs"
     }
 }
